requirements.txt
langchain-openai~=1.0.3
python-dotenv~=1.2.1
langchain~=1.0.7